requirements.txt
Django==5.2.9
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.27.2
requests==2.32.3
python-dotenv==1.0.1

# Optional: For advanced sentiment analysis
# openai==1.54.5
# nltk==3.9.1
# textblob==0.18.0

